Document that there is no amalgamation rule
The code provided by Gries and Hilpert does not do average amalgamation,
though the chapter suggests it should. It is not necessary here: instead
of merging points and replacing them with their average, we keep the
individual observations in each cluster and use the standard deviations
to decide which clusters to merge.